Add database schema and sample data scripts and made related to it entity changes.

+++ src/main/java/pl/edu/uksw/suncloud/weather/repository/MeasurementRepo.java
@@ -5,8 +5,8 @@ import pl.edu.uksw.suncloud.weather.domain.Measurement;
 
 import java.util.Date;
 
-public interface MeasurementRepo extends CrudRepository<Measurement, Integer> {
-    Measurement findById(Integer id);
+public interface MeasurementRepo extends CrudRepository<Measurement, Long> {
+    Measurement findById(Long id);
 
     Measurement findTop1ByOrderByDateTimeDesc();
     Measurement findByDateTime(Date date);
